requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
ansi2html==1.9.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
cachelib==0.9.0
cattrs==23.2.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dacite==1.8.1
dash==2.14.2
dash-bootstrap-components==1.5.0
dash-bootstrap-templates==1.1.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dill==0.3.7
discord.py==2.3.2
diskcache==5.6.3
exceptiongroup==1.2.0
Flask==3.0.0
Flask-Caching==2.1.0
frozenlist==1.4.1
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
multiprocess==0.70.15
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
pandas==2.1.4
platformdirs==4.1.0
plotly==5.18.0
pokemontcgsdk==3.4.0
psutil==5.9.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-cache==1.1.1
retrying==1.3.4
six==1.16.0
tenacity==8.2.3
typing_extensions==4.9.0
tzdata==2023.4
url-normalize==1.4.3
urllib3==2.1.0
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
